+++ lib/main.dart
@@ -16,4 +16,5 @@ class App extends StatelessWidget {
       home: new RootPage(),
     );
   }
+
 }